requirements.txt
astropy
h5py
hydra-core
matplotlib
pandas
torchaudio
torchvision
tqdm
wandb
zarr

